@@ -1,61 +1,99 @@
 {
   "name": "good-samaritan",
-  "description": "Find open issues from open source dependencies of your project.",
   "version": "1.6.1",
-  "author": "Scott Davis <stdavis@utah.gov>",
-  "bin": {
-    "good-samaritan": "./bin/run"
-  },
-  "bugs": "https://github.com/stdavis/good-samaritan/issues",
-  "dependencies": {
-    "@oclif/command": "^1.8.0",
-    "@oclif/config": "^1.17.0",
-    "@oclif/plugin-help": "^3.2.2",
-    "@octokit/plugin-retry": "^3.0.7",
-    "@octokit/plugin-throttling": "^3.4.1",
-    "@octokit/rest": "^18.1.0",
-    "chalk": "^4.1.0",
-    "cli-ux": "^5.5.1",
-    "node-fetch": "^2.6.1",
-    "package-json": "8.1.0",
-    "parse-github-url": "1.0.2",
-    "pretty-ms": "^7.0.1",
-    "read-pkg": "5.2.0"
-  },
-  "devDependencies": {
-    "@oclif/test": "^1.2.8",
-    "chai": "^4.3.0",
-    "debug": "^4.3.1",
-    "eslint": "^7.19.0",
-    "eslint-config-prettier": "^7.2.0",
-    "jest": "^26.6.3",
-    "msw": "^0.26.2",
-    "prettier": "^2.2.1",
-    "standard-version": "^9.1.0"
-  },
-  "engines": {
-    "node": ">=10.0.0"
-  },
-  "files": [
-    "/bin",
-    "/src"
-  ],
-  "homepage": "https://github.com/stdavis/good-samaritan",
+  "description": "Find open issues from open source dependencies of your project.",
   "keywords": [
     "cli"
   ],
-  "license": "MIT",
-  "main": "src/index.js",
-  "oclif": {
-    "bin": "good-samaritan"
-  },
+  "homepage": "https://github.com/stdavis/good-samaritan",
+  "bugs": "https://github.com/stdavis/good-samaritan/issues",
   "repository": "stdavis/good-samaritan",
-  "scripts": {
-    "posttest": "eslint .",
-    "test": "jest --watchAll --coverage --silent",
-    "release": "standard-version"
+  "license": "MIT",
+  "author": "Scott Davis <stdavis@utah.gov>",
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "bin": {
+    "good-samaritan": "./bin/run.js"
   },
-  "watch": {
-    "testrun": "src/*.js"
+  "files": [
+    "/bin",
+    "/dist",
+    "/oclif.manifest.json"
+  ],
+  "scripts": {
+    "build": "shx rm -rf dist && tsc -b",
+    "check": "tsc --noEmit",
+    "format": "prettier --write .",
+    "lint": "eslint .",
+    "prepack": "oclif manifest && oclif readme",
+    "postpack": "shx rm -f oclif.manifest.json",
+    "release": "standard-version",
+    "test": "vitest",
+    "posttest": "npm run lint"
+  },
+  "oclif": {
+    "additionalHelpFlags": [
+      "-h"
+    ],
+    "additionalVersionFlags": [
+      "-v"
+    ],
+    "bin": "good-samaritan",
+    "commands": {
+      "strategy": "single",
+      "target": "./dist/index.js"
+    },
+    "default": ".",
+    "plugins": [
+      "@oclif/plugin-help",
+      "@oclif/plugin-version"
+    ],
+    "topicSeparator": " "
+  },
+  "dependencies": {
+    "@oclif/core": "^4.0.27",
+    "@oclif/plugin-version": "^2.2.14",
+    "@octokit/plugin-retry": "^7.1.2",
+    "@octokit/plugin-throttling": "^9.3.2",
+    "@octokit/rest": "^21.0.2",
+    "chalk": "^5.3.0",
+    "cli-ux": "^6.0.9",
+    "conf": "^13.0.1",
+    "node-fetch": "^3.3.2",
+    "open": "^10.1.0",
+    "package-json": "10.0.1",
+    "parse-github-url": "1.0.3",
+    "pretty-ms": "^9.1.0",
+    "read-pkg": "9.0.1",
+    "un": "^0.0.0"
+  },
+  "devDependencies": {
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.12.0",
+    "@oclif/plugin-help": "^6.2.14",
+    "@oclif/test": "^4.0.9",
+    "@types/cli-progress": "^3.11.6",
+    "@types/debug": "^4.1.12",
+    "@types/node": "^22.7.5",
+    "@types/parse-github-url": "^1.0.3",
+    "chai": "^5.1.1",
+    "debug": "^4.3.7",
+    "eslint": "^9.12.0",
+    "eslint-config-prettier": "^9.1.0",
+    "globals": "^15.10.0",
+    "msw": "^2.4.9",
+    "prettier": "^3.3.3",
+    "prettier-plugin-organize-imports": "^4.1.0",
+    "prettier-plugin-packagejson": "^2.5.3",
+    "shx": "^0.3.4",
+    "standard-version": "^9.5.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.8.1",
+    "vitest": "^2.1.2"
+  },
+  "engines": {
+    "node": ">=18.0.0"
   }
 }
